backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
requests
opencv-python-headless